Bug 1914005: Move bounds checks out of call_indirect. r=rhunt
We currently do bounds checks for call_indirect at the masm level, and
always as 32-bit. This is not ideal now that we have table64, since it
forces us to clamp our address and length values down to 32 bits.

This patch removes the bounds check from masm and the corresponding
optimization logic from lowering. Instead, it moves the optimization
logic into MIR and LIR. Specifically:

- If the index is constant and less than the table's min size, the
  bounds check is eliminated. (This will be done in the next patch.)
- If the table has a fixed length, a constant is emitted instead of a
  load, and we will compare against an immediate instead of a register
  or memory operand.
- If the bounds check limit is an MWasmLoadInstance (as it is for
  tables), then the load will be emitted at use as a memory operand to
  the bounds check instead of a load to a register.

Differential Revision: https://phabricator.services.mozilla.com/D253576

+++ js/src/wasm/WasmInstanceData.h
@@ -166,8 +166,10 @@ struct MemoryInstanceData {
 // to bounds-check and index the table.
 
 struct TableInstanceData {
-  // Length of the table in number of elements (not bytes).
-  uint32_t length;
+  // Length of the table in number of elements (not bytes). Although the type is
+  // uint64_t, the maximum value fits in 32 bits -- this value can safely be
+  // loaded as either a 32-bit value or a 64-bit value.
+  uint64_t length;
 
   // Pointer to the array of elements (which can have various representations).
   // For tables of anyref this is null.
